fix 3 step docker restart for flannel
remove restart docker handler from minion.  it's unused.

notifications are run in order "per task" so the main task notified
restart docker and restart kubelet.

restart docker notified delete docker0 and start docker.

but since that is considered a different task, the next handler run was
restart kubelet, which requires docker to be up.  to fix, we need to
call all 3 docker restart steps in the main task, instead of chaining
them.